refac: style 추출 및 컴포넌트 이름 수정
