+++ test-web/src/components/DataListView/PASingle.js
@@ -1,5 +1,6 @@
 import { useState, useEffect } from 'react';
 import DataList from './DataList';
+import style from './style/pasinglestyle';
 
 
 // 데이터 목록 조회 페이지 컴포넌트
@@ -30,27 +31,3 @@ const PASingle = ({ startDate, endDate, data }) => {
 };
 
 export default PASingle;
-
-const style = {
-  listContainer: {
-    textAlign: 'center',
-    width: '100%',
-    paddingRight: '0px',
-    paddingBottom: '0',
-    height: 'auto',
-  },
-  paginationBar: {
-    marginTop: '40px',
-    width: '100%',
-    justifyContent: 'center',
-  },
-  paginationContainer: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-  },
-  formControl: {
-    minWidth: 120,
-    marginLeft: '20px',
-  },
-};
