+++ test-web/src/components/DataListView/style/pasinglestyle.js
@@ -1,0 +1,25 @@
+const style = {
+    listContainer: {
+      textAlign: 'center',
+      width: '100%',
+      paddingRight: '0px',
+      paddingBottom: '0',
+      height: 'auto',
+    },
+    paginationBar: {
+      marginTop: '40px',
+      width: '100%',
+      justifyContent: 'center',
+    },
+    paginationContainer: {
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'space-between',
+    },
+    formControl: {
+      minWidth: 120,
+      marginLeft: '20px',
+    },
+  };
+
+  export default style
