+++ test-web/src/components/DataListView/style/rejecteddatalistcompstyle.js
@@ -1,0 +1,39 @@
+const navy = '#0F3659';
+
+const style = {
+    listContainer: {
+      textAlign: 'center',
+      width: '100%',
+      paddingRight: '0px',
+      paddingBottom: '0',
+      height: 'auto',
+    },
+    paginationBar: {
+      marginTop: '40px',
+      width: '100%',
+      justifyContent: 'center',
+    },
+    deletBtnContainer: {
+      display: 'flex',
+      margin: '20px 0',
+      padding: '0px 100px',
+      width: '100%',
+      justifyContent: 'start',
+    },
+    deleteBtn: {
+      backgroundColor: 'transparent',
+      color: navy,
+      fontWeight: '600',
+    },
+    formControl: {
+      minWidth: 120,
+      marginLeft: '20px',
+    },
+    paginationContainer: {
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'space-between',
+    },
+  };
+
+  export default style
